Separacao das paginas originais
